requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-autocomplete-light==3.8.2
django-bootstrap-form==3.4
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
